backend/requirements.txt
cryptography
paramiko
flask
flask-cors
requests
python-dotenv
tenacity
pyodbc
eth-account
base58
solders
web3
apscheduler
pytest
boto3
awscli
werkzeug
solana